[nCoop]-005
 - add confirm delete /w sweetalert
 - close modal and reload data after insert
 - pagination all page

+++ public/js/controllers/documentCtrl.js
@@ -1,26 +1,48 @@
 documentModule.controller('documentCtrl', ['$scope','$uibModal','$log', 'DocumentsService', function($scope,$uibModal,$log, DocumentsService) {
     $scope.loading = true;
+    
     $scope.documents = []
         ,$scope.currentPage = 1
         ,$scope.numPerPage = 10
         ,$scope.maxSize = 5;
 
-    $scope.items = ["AA","BB","CC","DD","EE",];
 
-    DocumentsService.get()
-        .success(function(data) {
+    var getDocument = function () {
+        DocumentsService.get().success(function(data) {
             console.log("Documents data retrieving success.");
             console.log(data);
             $scope.documents = data;
             $scope.loading = false;
         });
+    }
+
+    getDocument();
 
     $scope.deleteDocument = function(id) {
 
-        $scope.loading = true;
-        DocumentsService.delete(id).success(function(data) {
-            $scope.documents = data;
-            $scope.loading = false;
+        //sweet alert here
+        swal({
+            title: "คุณแน่ใจหรือ?",
+            text: "การกระทำนี้ไม่สามารถกู้ข้อมูลคืนได้",
+            type: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: "ใช่, ลบเอกสารนี้!",
+            cancelButtonText: "ยกเลิก",
+            closeOnConfirm: false,
+            closeOnCancel:false,
+            html: false
+        }, function(isConfirm){
+            if(isConfirm) {
+                $scope.loading = true;
+                DocumentsService.delete(id).success(function(data) {
+                    $scope.documents = data;
+                    $scope.loading = false;
+                });
+                swal("ลบ!","เอกสารนี้ถูกลบออกแล้ว","success");
+            } else {
+                swal("ยกเลิก", " ","error");
+            }
         });
     };
 
@@ -33,8 +55,8 @@ documentModule.controller('documentCtrl', ['$scope','$uibModal','$log', 'Documen
         });
 
         modalInstance.result.then(function() {
-        }, function() {
             $log.info('Modal dismissed at: ' + new Date());
+            getDocument();
         });
     }
